Update: Get example running with adding categories
Reorganizes code a bit and gets an example running with adding random circles every 2.5 seconds.
