+++ app/src/index.css
@@ -1,3 +1,9 @@
+*,
+*:before,
+*:after {
+  box-sizing: content-box;
+}
+
 body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
@@ -5,9 +11,11 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+  background-color: #acd;
 }
 
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+svg {
+  margin: 0;
+  max-width: 100%;
+  max-height: 100%;
 }
